+++ frontend/package.json
@@ -10,6 +10,7 @@
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
     "react-scripts": "^5.0.1",
     "tailwindcss": "^3.4.1",
     "web-vitals": "^2.1.4"
